+++ EchoLedger/Features/Transaction/Domain/UseCases/AddTransactionInput.swift
@@ -12,7 +12,7 @@ struct AddTransactionInput {
     let userId: UUID
     let label: String
     let date: Date
-    let totalAmount: Decimal
+    let totalAmount: Double
     let note: String?
     let isExpense: Bool
     let category: TransactionCategory
