+++ EchoLedgerTests/Features/Transaction/UseCases/GetTransactionsByDateRangeTests.swift
@@ -35,7 +35,7 @@ final class GetTransactionsByDateRangeTests: XCTestCase {
     }
 
     /// Seeds a transaction at a given date belonging to the shared userId.
-    private func seedTransaction(date: Date, amount: Decimal = 30) async throws {
+    private func seedTransaction(date: Date, amount: Double = 30) async throws {
         let split = await TransactionSplit(accountId: UUID(), amount: amount)
         let transaction = await Transaction(userId: userId,
                                             label: "Test",
